refactor : moved ArticleInterface's location and applied
