--- src/interface/article.ts
@@ -1,4 +1,0 @@
-export interface ArticleInterface {
-  title: string
-  content: string
-}
